Move item stuff to it's own controller
